release v1.0

+++ internal/model/entity/message_entity.go
@@ -5,7 +5,7 @@ import "database/sql"
 type MessageEntity struct {
 	Id        string
 	Message   sql.NullString
-	TopicId   int
+	TopicId   int64
 	CreatedAt int64
 	UpdatedAt int64
 }
